refactor: move components to closer folders
