+++ src/app/life/components/divider-bar.tsx
