+++ src/app/life/components/life-link.tsx
