+++ src/app/life/components/life-sub-topic.tsx
