+++ src/app/life/components/life-topic.tsx
