+++ src/app/life/components/nav-cards.tsx
@@ -25,7 +25,7 @@ const NavCardInner = ({ page }: { page: NavPageType }): React.JSX.Element => (
   </>
 );
 
-const NavCards = ({ pages }: { pages: NavPageType[] }): React.JSX.Element => {
+export const NavCards = ({ pages }: { pages: NavPageType[] }): React.JSX.Element => {
   return (
     <div className="flex flex-wrap justify-center gap-10">
       {pages.map((page) => {
@@ -48,5 +48,3 @@ const NavCards = ({ pages }: { pages: NavPageType[] }): React.JSX.Element => {
     </div>
   );
 };
-
-export default NavCards;
