+++ src/app/life/components/tldr.tsx
