+++ src/app/life/council/activities/page.tsx
@@ -1,5 +1,5 @@
-import LifeEvent from '@/components/life/life-activities-event';
-import LifeTopic from '@/components/life/life-topic';
+import LifeTopic from '@/app/life/components/life-topic';
+import Event from './components/event';
 import lifeEvents from './life-council-activities.json' assert { type: 'json' };
 
 interface LifeEventItem {
@@ -15,7 +15,7 @@ const Page = (): React.JSX.Element => {
     <div className="xl:ml-8 xl:max-w-[min(56rem,73%)] lg:max-w-4xl">
       <LifeTopic topic="系上活動" />
       {typedEvents.map((event, idx) => (
-        <LifeEvent
+        <Event
           key={idx}
           title={event.title}
           description={event.description}
