+++ src/app/life/course/[id]/components/list.tsx
