+++ src/app/life/course/[id]/components/table.tsx
