+++ src/app/life/information/internet/page.tsx
@@ -1,11 +1,11 @@
 'use client';
 import React, { useEffect, useState } from 'react';
-import DividerBar from '@/components/divider-bar';
-import LifeCourseContent from '@/components/life/life-course-content';
-import InformTable from '@/components/life/life-information-table';
-import LifeLink from '@/components/life/life-link';
-import TLDR from '@/components/life/life-tldr';
-import LifeTopic from '@/components/life/life-topic';
+import LifeCourseContent from '@/app/life/components/content-block';
+import DividerBar from '@/app/life/components/divider-bar';
+import LifeLink from '@/app/life/components/life-link';
+import LifeTopic from '@/app/life/components/life-topic';
+import TLDR from '@/app/life/components/tldr';
+import InformationTable from '@/app/life/information/components/information-table';
 import contentData from './content.json';
 
 interface LinkItem {
@@ -55,7 +55,7 @@ const Page = (): React.JSX.Element => {
             {item.content && <LifeCourseContent content={item.content} />}
             {item.table && (
               <div className="mt-4">
-                <InformTable table={item.table} />
+                <InformationTable table={item.table} />
               </div>
             )}
             {item.tldr && (
