+++ src/app/life/interview/[professorName]/components/life-interview-answer.tsx
