+++ src/utils/search-redirect.ts
